Using search samples from the sample view and refactoring to move utility methods from termsetting/samplelst to groups
