+++ client/termsetting/handlers/samplelst.ts
@@ -101,105 +101,3 @@ export function fillTW(tw: SampleLstTW) {
 		}
 	}
 }
-
-export function getSamplelstTW(groups: any, name = 'groups', notIn = true) {
-	let disabled = true
-	const values = {}
-	const qgroups: any = []
-	let samples: any
-	for (const group of groups) {
-		values[group.name] = { key: group.name, label: group.name, color: group.color }
-		samples = getGroupSamples(group)
-		const qgroup = {
-			name: group.name,
-			in: true,
-			values: samples
-		}
-		qgroups.push(qgroup)
-	}
-	if (groups.length == 1 && notIn) {
-		const name2 = 'Not in ' + groups[0].name
-		values[name2] = { key: name2, label: name2, color: '#aaa' }
-		qgroups.push({
-			name: name2,
-			in: false,
-			values: samples
-		})
-	}
-	const tw = {
-		isAtomic: true,
-		term: { name, type: 'samplelst', values },
-		q: {
-			groups: qgroups,
-			groupsetting: { disabled }
-		}
-	}
-	return tw
-
-	function getGroupSamples(group: any) {
-		const values: any = []
-		for (const item of group.items) {
-			const value: { sampleId: string; sample?: string } = { sampleId: item.sampleId }
-			if ('sample' in item) {
-				disabled = false
-				value.sample = item.sample
-			}
-			values.push(value)
-		}
-		return values
-	}
-}
-
-export function getSamplelstTWFromIds(ids: number[]) {
-	const qgroups: any = []
-	let samples: any
-	const name = 'group'
-	const values = ids.map(id => {
-		return { sampleId: id }
-	})
-	const qgroup = {
-		name: name,
-		in: true,
-		values
-	}
-
-	const tw = {
-		isAtomic: true,
-		term: { name, type: 'samplelst', values: { [name]: { key: name, list: values } } },
-		q: {
-			groups: [qgroup]
-		}
-	}
-	return tw
-}
-
-export function getSampleFilter(sampleId: number) {
-	const group = { name: '', index: 0, items: [{ sampleId }] }
-	const samplelst = getSamplelstTW([group], '', false)
-	const filter = getFilter(samplelst)
-	return filter
-}
-
-export function getFilter(samplelstTW: any) {
-	let i = 0
-	let noEdit = true
-	for (const field in samplelstTW.term.values) {
-		const values = samplelstTW.q.groups[i].values
-		samplelstTW.term.values[field].list = values
-		if (values[0] && 'sample' in values[0]) noEdit = false
-		i++
-	}
-	const filter = {
-		type: 'tvslst',
-		in: true,
-		join: '',
-		lst: [
-			{
-				type: 'tvs',
-				tvs: { term: samplelstTW.term },
-				noEdit
-			}
-		]
-	}
-	return filter
-}
